refactor: make extensions to cogs, fix: music player queue error
