+++ .env.example
@@ -1,0 +1,1 @@
+DISCORD_TOKEN =
