+++ .gitignore
@@ -1,0 +1,4 @@
+.env
+*.json
+
+__pycache__
