+++ config/const.py
@@ -1,0 +1,2 @@
+# 等級系統設定
+LEVEL_MULTIPLIER = 25  # 每級所需經驗值倍數
