+++ config/default_json.py
@@ -1,0 +1,134 @@
+user_data = {}
+
+shop_items = [
+    {
+        "id": "title_1",
+        "name": "初學者稱號",
+        "description": "新手專屬稱號",
+        "price": 100,
+        "type": "title",
+    },
+    {
+        "id": "title_2",
+        "name": "黃金會員",
+        "description": "尊貴黃金會員稱號",
+        "price": 500,
+        "type": "title",
+    },
+    {
+        "id": "bg_1",
+        "name": "星空背景",
+        "description": "個人資料星空背景",
+        "price": 300,
+        "type": "background",
+    },
+    {
+        "id": "role_color",
+        "name": "自訂角色顏色",
+        "description": "解鎖自訂角色顏色權限",
+        "price": 800,
+        "type": "perk",
+    },
+    {
+        "id": "double_xp",
+        "name": "雙倍經驗卡",
+        "description": "24小時內獲得雙倍經驗",
+        "price": 200,
+        "type": "boost",
+    },
+    {
+        "id": "fortune_boost",
+        "name": "幸運水晶",
+        "description": "提升占卜結果品質",
+        "price": 150,
+        "type": "fortune",
+    },
+]
+
+
+tarot_cards = [
+    {
+        "name": "愚者",
+        "meaning": "新的開始、冒險精神",
+        "type": "major",
+        "image": "🃏",
+    },
+    {
+        "name": "魔術師",
+        "meaning": "創造力、技能",
+        "type": "major",
+        "image": "🧙",
+    },
+    {
+        "name": "女祭司",
+        "meaning": "直覺、神秘知識",
+        "type": "major",
+        "image": "🔮",
+    },
+    {"name": "女皇", "meaning": "豐饒、母性", "type": "major", "image": "👑"},
+    {"name": "皇帝", "meaning": "權威、結構", "type": "major", "image": "👨‍💼"},
+    {"name": "戰車", "meaning": "意志力、勝利", "type": "major", "image": "🛡️"},
+    {
+        "name": "力量",
+        "meaning": "勇氣、內在力量",
+        "type": "major",
+        "image": "💪",
+    },
+    {
+        "name": "隱者",
+        "meaning": "內省、尋求真理",
+        "type": "major",
+        "image": "🧓",
+    },
+    {
+        "name": "命運之輪",
+        "meaning": "命運轉變、契機",
+        "type": "major",
+        "image": "🔄",
+    },
+    {"name": "正義", "meaning": "公平、決定", "type": "major", "image": "⚖️"},
+    {
+        "name": "倒吊人",
+        "meaning": "犧牲、新視角",
+        "type": "major",
+        "image": "🙃",
+    },
+    {"name": "死神", "meaning": "結束、轉變", "type": "major", "image": "💀"},
+    {"name": "節制", "meaning": "平衡、調和", "type": "major", "image": "⚗️"},
+    {
+        "name": "惡魔",
+        "meaning": "束縛、物質主義",
+        "type": "major",
+        "image": "😈",
+    },
+    {"name": "塔", "meaning": "突變、啟示", "type": "major", "image": "⚡"},
+    {"name": "星星", "meaning": "希望、靈感", "type": "major", "image": "⭐"},
+    {"name": "月亮", "meaning": "幻覺、潛意識", "type": "major", "image": "🌙"},
+    {"name": "太陽", "meaning": "成功、喜悅", "type": "major", "image": "☀️"},
+    {"name": "審判", "meaning": "覺醒、重生", "type": "major", "image": "👼"},
+    {"name": "世界", "meaning": "完成、成就", "type": "major", "image": "🌍"},
+    {
+        "name": "寶劍Ace",
+        "meaning": "新的想法、突破",
+        "type": "minor",
+        "image": "⚔️",
+    },
+    {
+        "name": "權杖Ace",
+        "meaning": "新的開始、能量",
+        "type": "minor",
+        "image": "🔥",
+    },
+    {
+        "name": "聖杯Ace",
+        "meaning": "新的情感、直覺",
+        "type": "minor",
+        "image": "💧",
+    },
+    {
+        "name": "錢幣Ace",
+        "meaning": "新的財富、機會",
+        "type": "minor",
+        "image": "💰",
+    },
+]
