+++ core/classes.py
@@ -1,0 +1,6 @@
+from discord.ext import commands
+
+
+class Cog(commands.Cog):
+    def __init__(self, bot):
+        self.bot = bot
